% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_categorical.R
\name{categorical_tbl}
\alias{categorical_tbl}
\title{Title}
\usage{
categorical_tbl(
  df,
  grouping_var = NULL,
  header = TRUE,
  total = TRUE,
  p_value = TRUE
)
}
\arguments{
\item{total}{}
}
\value{

}
\description{
Title
}
