% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_continuous.R
\name{continuous_tbl}
\alias{continuous_tbl}
\title{Title}
\usage{
continuous_tbl(
  df,
  grouping_var = NULL,
  digit = 2,
  total = TRUE,
  header = TRUE,
  modification = "",
  p_value = TRUE
)
}
\arguments{
\item{header}{}
}
\value{

}
\description{
Title
}
