% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_goat.R
\name{tbl_goat}
\alias{tbl_goat}
\title{Title}
\usage{
tbl_goat(
  df,
  grouping_var = NULL,
  digit = 2,
  p_value = TRUE,
  total = TRUE,
  header = TRUE,
  kabel = TRUE
)
}
\arguments{
\item{kabel}{}
}
\value{

}
\description{
Title
}
